修复了页面传参刷新后丢失的bug
